[starvation] make thread domain more precise
Summary:
Add precision to analysis by elaborating the thread-status domain. This is done by having unknown (bottom), UI, BG or Any (both/top) elements in the lattice.  This way, when we branch on thread-identity (if I am on UI thread do this, otherwise do that), we know that in one branch we are on UI thread and on the  other we are *not* on the UI thread (BG thread), where previously the other branch would just go to top.

With this knowledge we can throw away pairs that come from callees which run on a thread that is impossible, given the current caller thread identity.  This can happen when annotations are used incorrectly, and since this is the purview of annot-reachability, we just drop those pairs entirely.

Reviewed By: skcho

Differential Revision: D18202175

fbshipit-source-id: be604054e