[starvation] make report-map per file
Summary:
The upcoming whole-program analysis will need to log reports in different source files simultaneously.
Here, the data structure containing the reports, and in charge of deduplication is generalised to
a map from source file to what it was previously.  This way, writing out the reports is possible even
with multiple source files.

Reviewed By: skcho

Differential Revision: D18394994

fbshipit-source-id: 5f5ecc27c