[call-graph scheduling] fix bug in node removal
Summary:
The order was wrong: the map from procnames to node-ids was cleaned first, but to clean the map from node-ids to nodes, we need the id and we have already removed it.

The symptom was that effectively no new leaves are created by "removing" a node, so the call graph scheduler quickly devolves to the file-based one.

Reviewed By: skcho

Differential Revision: D18448209

fbshipit-source-id: f272a8112