[inferbo] Substitution of array block of default case in Java
Summary:
There was a precision loss during the substitution of array block.  For example:

Callee's abstract memory includes an array block as follows, where `a` is a parameter.

```
a.elements -> { a.elements[*] with a.elements.size }
```

Callers' abstract memory includes a pointer that may point to multiple array blocks.

```
c -> { x, y }
x.elements -> { x.elements[*] with x.elements.size }
y.elements -> { y.elements[*] with y.elements.size }
```

When the callee is called with the parameter `c`, the callees memory is substituted to:

```
x.elements -> { x.elements[*] with top , y.elements[*] with top }
y.elements -> { x.elements[*] with top , y.elements[*] with top }
```

because `a.elements[*]` was substituted to `{ x.elements[*] , y.elements[*] }`
and `a.elements.size` was substituted to `top ( = x.elements.size join y.elements.size )`.

This diff tries to keep the precision in the specific case, not to join the sizes of array blocks.
So now the same callee's abstract memory is substituted to:

```
x.elements -> { x.elements[*] with x.elements.size }
y.elements -> { y.elements[*] with y.elements.size }
```

Reviewed By: ngorogiannis

Differential Revision: D18480585

fbshipit-source-id: b70e63c22