[self in block] Add a check for strongSelf not checked for null
Summary:
The variable strongSelf, because it is equal to a weak captured pointer, needs to be checked for null before being used, otherwise it could be null by the time the block is executed.

Added this check to the SelfInBlock checker, and removed it from biabduction.

We want to migrate all the objc checks from biabduction, so it will be easier to change and faster and more reliable. Moreover, this check is more general, it will flag any use of unchecked strongSelf, not just a dereference.

Reviewed By: skcho

Differential Revision: D18403849

fbshipit-source-id: a9cf5d80b