[nullsafe] Changes in wording of error messages: unified and more clear language
Summary:
Whole bunch of changes aimed to make error messages more clear and
concise.

1/ Wording and language is unified. We make errors sound more like a
type system violations, rather than linter reccomendations.

Particularly, we refrain from saying things like "may be null" - this is
a linter-style statement that may provoke discussions (what if the
developer knows it can not be null in this particular case).

Instead, we refer to declared nullability and nullability of actual values. This way, it is more clear that this is not a heuristic, this is how rules of a type-system work.

2/ Additionally, we drop things like field class in places when the
context should be clear by who looks at the error. We expect the user
sees the code and the error caption. So e.g. we don't repeat the word "field"
twice.

3/ In cases when we are able to retrieve formal param name, we include it for
usability.

4/ For Field not initialized error, we refer to Initializer methods:
this is a non-obvious but important nullsafe feature.

Reviewed By: artempyanykh

Differential Revision: D18569762

fbshipit-source-id: 9221d7102