[build] fix opam_retry
Summary:
Not sure what's going on but even though `set -e` is passed,
`install_opam_deps` was seen as succeeding even when the `opam install`
command failed. This prevented `opam_retry` from getting triggered (and
the error was also silently ignored).

Also get rid of a couple of useless subshells in `opam_retry`.

Reviewed By: jberdine

Differential Revision: D18615860

fbshipit-source-id: 360232623