[infer] Record the value of function arguments in ProcnameDispatcher calls
Summary: Preperation diff to use `ProcnameDispatcher` for Pulse: it changes function arguments, i.e. `ProcnameDispatcher.Call.FuncArg`, to a record in order to track the value of arguments. To do that, it changes `ProcnameDispatcher.Call` into a functor so that we can parametrize over the type of the value without making changes upwards.

Reviewed By: jvillard

Differential Revision: D18590224

fbshipit-source-id: 6a13fbc1a