[infer] Remove marker from procname dispatcher
Summary: This diff removes `'markers` and `'captured_types` from the procname dispatcher. They are for checking an integrity when a type is captured from template parameters then it is used to match in parameters. However, we have not used that feature, so which simply complicates the types in the dispatcher without any gain at the moment.

Reviewed By: jvillard

Differential Revision: D18706254

fbshipit-source-id: f493778d7