[pulse] havoc formals passed by reference to unknown procedures
Summary:
This gets rid of false positives when something invalid (eg null) is
passed by reference to an initialisation function. Havoc'ing what the
contents of the pointer to results in being optimistic about said
contents in the future.

Also surprisingly gets rid of some FNs (which means it can also
introduce FPs) in the `std::atomic` tests because a path condition
becomes feasible with havoc'ing.

There's a slight refinement possible where we don't havoc pointers to
const but that's more involved and left as future work.

Reviewed By: skcho

Differential Revision: D18726203

fbshipit-source-id: 264b5daeb