[self-in-block] Add new issue type MULTIPLE_WEAKSELF
Summary: Raises an error when weakSelf is used multiple times in a block. The issue is that there could be unexpected behaviour. Even if `weakSelf` is not nil in the first use, it could be nil in the following uses since the object that `weakSelf` points to could be freed anytime.

Reviewed By: skcho

Differential Revision: D18765493

fbshipit-source-id: 37fc652e3