[starvation] fix FPs due to mishandling wait
Summary:
`Object.wait()` must be called on a locked monitor and it releases the
lock immediately, as far as other threads are concerned
(it also magically re-takes the lock when the monitor is `notified`).
Starvation can only occur if the UI thread is waiting
a lock that is distinct to that being waited on.

The check present was over-approximate in that it was checking that there exists a lock held by the UI thread and the thread issuing the `wait`, but did not make sure that lock was *not* the one waited on.

Amusingly, the e2e test was correct, but the reporting code wasn't.

Reviewed By: dulmarod

Differential Revision: D18782919

fbshipit-source-id: b3b98239e