[litho] Change the domain
Summary:
Making `MethodCalled` an inverted map from created location to method calls results in not being able to track a builder that is created in two different branches of a conditional with different types. Instead, we can make `MethodCalled` simply a map and also change  `Created` to be a map from access paths to a set of created locations.

To deal with the case of setting a prop only in one branch, we need to ensure that whenever we call a create method, we add a binding to `MethodCalled` with an empty list of methods so that its intersection with a non-empty one is empty.

Reviewed By: skcho

Differential Revision: D18883097

fbshipit-source-id: b3464ca20