[typ][fieldname] eliminate uses of Java.from_string
Summary: This function allows any string, and in particular empty class names.  As a first step eliminate it in favour of a function that forces the caller to specify distinct class and field names.  It turns out that the frontend already has them, so it saves effort along the way.

Reviewed By: jvillard

Differential Revision: D18953136

fbshipit-source-id: ff3cdfda5