[typ][fieldname] simplify and streamline interface
Summary:
Remove Clang and Java submodules of Typ.Fieldname. They are unnecessary and they reflect a fake dichotomy: there is only one fieldname type.  To distinguish between fields of Java classes and other C constructs, there is a helper function provided, but the idea is simple: obtain the class type the field belongs to, and check if it's a Java class.

This diff still preserves behaviour, but removes as many functions as possible from the interface, to leave a small surface.

Reviewed By: mityal

Differential Revision: D18962423

fbshipit-source-id: ffe6933ee