[topl] Generate simpler monitor for deterministic states.
Summary:
This is an optimization. We ask the user to tell us which states are nondeterministic, and we
generate code that handle nondeterminism only for those states. It is common for only one state per
TOPL property to be nondeterministic. This speeds up the biabduction-analysis of the monitor by a
factor of ~10. But, using the monitor is only a little faster.

Facebook

Reviewed By: jvillard

Differential Revision: D19160286

fbshipit-source-id: 4dd39769a