Cleaning in the Java Frontend
Summary:
From a set of classes [classes], JClasspath.load_program builds a
program. This program contains (implicitely) the set of active classes
and we should not keep using the set [classes] and the generated
program as the same time, because they may not be synched
anymore. (In particular, in a forthcoming diff load_program may add
new classes during program construction)

This patch forces this discipline. It adds a [in_classes]
query for programs and makes sure we use it everywhere we should.

Reviewed By: ngorogiannis

Differential Revision: D18833335

fbshipit-source-id: a522f320c