[sledge] Replace Memory and Concat Term constructors with eq_concat
Summary:
The exposed constructors for Memory and Concat Terms are only used in
a very special idiom: to construct an equality between a single Memory
chunk and the Concat of multiple Memory chunks. This diff specializes
and simplifies by exposing a Term.eq_concat constructor for this
idiom, and removes the underlying Term.memory and Term.concat
constructors.

Reviewed By: ngorogiannis

Differential Revision: D19221866

fbshipit-source-id: 4842737d2