[litho] Make analysis null return aware
Summary:
This diff avoids that null-retuned path's abstract value ruins that of non-null-returned path.
What this diff does is: when joining two abstract states, one is null-return-path and the other is
non-null-return-path (`return obj;`), it keeps the method calls of `obj` from the
non-null-return-path.

While this design is unsound, I think it should work in practice.

Reviewed By: ezgicicek

Differential Revision: D19348313

fbshipit-source-id: cf5d0f3ff