[impurity] Rely on set of skipped functions to determine impurity
Summary:
Currently, impurity analysis is oblivious to skipped functions which might e.g. return a non-deterministic value, write to memory or have some other side-effect. This diff fixes that by relying on Pulse's skipped functions to determine impurity. Any unknown function which is not modeled to be pure is assumed to be impure.

This is a heuristic. We could have assumed them to be pure by default as well.

Reviewed By: jvillard

Differential Revision: D19428514

fbshipit-source-id: 82efe04f9